+++ lib/just_tooltip.dart
@@ -2,5 +2,6 @@ export 'src/enums.dart';
 export 'src/just_tooltip_controller.dart';
 export 'src/just_tooltip.dart';
 export 'src/just_tooltip_overlay.dart';
+export 'src/just_tooltip_theme.dart';
 export 'src/tooltip_position_utils.dart';
 export 'src/tooltip_shape_painter.dart';